requirements.txt
bs4
docopt-ng
epicstuff
lxml
requests